% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asmice.R
\name{as.mice}
\alias{as.mice}
\title{Converts an multiply imputed dataset obtained from DURR and IURR functions into a \code{mids} object}
\usage{
as.mice(data, imp)
}
\arguments{
\item{data}{The original incomplete data set}

\item{imp}{The return value from \code{GMdurr} or \code{GMiurr} with the option of \code{included=FALSE}}
}
\value{
An object of class \code{\link{mids}}
}
\description{
This function converts imputed data obtained from \code{GMdurr} or \code{GMiurr} into an object of class \code{mids}. The original incomplete data set needs to be available so that we know where the missing data are. This function is useful to manipulate the imputed data sets with the help of functions from \code{mice} package.
}
\examples{
data(GMdata)
GM.imp=GMdurr(data=GMdata)
imp=as.mice(data=GMdata,imp=GM.imp)
}

