% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asmice.R
\name{GMdata}
\alias{GMdata}
\title{High-dimensional data set with general missing patterns.}
\format{A data frame with 100 rows and 200 variables}
\description{
A data set containing 100 observations of 200 continuous variables (\code{z_1} to \code{z_200}). The first three variables have missing values with numbers of 28, 30 and 28, respectively.
}
\examples{
data(GMdata)
}

