% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMdurr.R
\name{GMdurr}
\alias{GMdurr}
\title{Multiple Imputation through Direct Use of Regularized Regression(DURR) in the presence of General missing patterns}
\usage{
GMdurr(
  data,
  m = 5,
  family = rep("gaussian", 3),
  method = c("lasso"),
  included = FALSE,
  burn = 20
)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data with general missing patterns. Missing values are coded as NA.}

\item{m}{Number of multiple imputations. The default is \code{m=5}.}

\item{family}{Specify the type of variables that have missing values. Can be either a single string or a vector of string with length of the number of variables having missing values. Each string can be "gaussian", or "binary", or "poisson".}

\item{method}{The method to be used in regularized regression. Can be "lasso", "elastic net", and "adaptive lasso".}

\item{included}{If \code{TRUE}, GMdurr will include all observed values among m complete datasets.}

\item{burn}{A scalar giving the number of burns for the convergence of iterations. The default is \code{20}.}
}
\value{
imp If \code{included=FALSE}, a list of components with the generated multiple imputations. The number of components equals the number of variables in the data set that have missing values. Each part of the list is a nmis[j] by m matrix of imputed values for variable j.

imp.included If \code{included=TRUE}, a list of components with the generated multiple imputations. The number of components equals the number of variables in the data set that have missing values. Each part of the list is a nrow(data) by m matrix of imputed values and observed values for variable j.
}
\description{
Multiple Imputation through Direct Use of Regularized Regression(DURR) in the presence of General missing patterns
}
\examples{
data(GMdata)
GMdurr(data=GMdata)
}
