% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0927MIdurr.R
\name{MIdurr}
\alias{MIdurr}
\title{Multiple Imputation through Direct Use of Regularized Regression(DURR) for univariate missing data pattern}
\usage{
MIdurr(data, family = c("gaussian"), method = c("lasso"), m = 5)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data. Missing values are coded as NA.}

\item{family}{Specify the type of the missing data. Can be "gaussian", or "binary", or "poisson".}

\item{method}{The method to be used in regularized regression. Can be "lasso", "elastic net","adaptive lasso" or "blasso".}

\item{m}{Number of multiple imputations. The default is m=5.}
}
\value{
n.missing Number of missing values in the data.

col.missing Which column contains the missing values.

impute The imputations(column number equals to m). Each column corresponds to one imputation.
}
\description{
Multiple Imputation through Direct Use of Regularized Regression(DURR) for univariate missing data pattern
}
\examples{
data(datagaussian)
data(databinary)
data(datapoisson)
MIdurr(data=datagaussian,method="lasso",family="gaussian")
MIdurr(data=datagaussian,method="blasso",family="gaussian")
MIdurr(data=databinary,method="elastic net",family="binary")
MIdurr(data=databinary,method="adaptive lasso",family="binary")
MIdurr(data=datapoisson,method="elastic net",family="poisson")
MIdurr(data=datapoisson,method="adaptive lasso",family="poisson")
}
