% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0927MIiurr.R
\name{MIiurr}
\alias{MIiurr}
\title{Multiple Imputation through Indirect Use of Regularized Regression(IURR) for univariate missing data pattern}
\usage{
MIiurr(data, family = c("gaussian"), method = c("lasso"), m = 5)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data. Missing values are coded as NA.}

\item{family}{Specify the type of the missing data. Can be "gaussian", or "binary", or "poisson".}

\item{method}{The method to be used in regularized regression. Can be "lasso", "elastic net","adaptive lasso" or "blasso".}

\item{m}{Number of multiple imputations. The default is m=5.}
}
\value{
n.missing Number of missing values in the data.

col.missing Which column contains the missing values.

impute The imputations(Column number equals to m). Each column corresponds to one imputation.
}
\description{
Multiple Imputation through Indirect Use of Regularized Regression(IURR) for univariate missing data pattern
}
\examples{
data(datagaussian)
data(databinary)
data(datapoisson)
MIiurr(data=datagaussian,method="lasso",family="gaussian")
MIiurr(data=datagaussian,method="blasso",family="gaussian")
MIiurr(data=databinary,method="elastic net",family="binary")
MIiurr(data=databinary,method="adaptive lasso",family="binary")
MIiurr(data=datapoisson,method="elastic net",family="poisson")
MIiurr(data=datapoisson,method="adaptive lasso",family="poisson")
}

